feat: add typescript to sandboxes
