+++ sandboxes/doughnut/src/components/doughnutChart.ts
@@ -1,27 +1,27 @@
-<script>
-import { defineComponent, h } from 'vue'
-import { Bar } from 'vue-chartjs'
+import { defineComponent, h, PropType } from 'vue'
+
+import { Doughnut } from 'vue-chartjs'
 import {
   Chart as ChartJS,
   Title,
   Tooltip,
   Legend,
-  BarElement,
+  ArcElement,
   CategoryScale,
-  LinearScale
+  PluginOptionsByType
 } from 'chart.js'
 
-ChartJS.register(Title, Tooltip, Legend, BarElement, CategoryScale, LinearScale)
+ChartJS.register(Title, Tooltip, Legend, ArcElement, CategoryScale)
 
 export default defineComponent({
-  name: 'BarChart',
+  name: 'DoughnutChart',
   components: {
-    Bar
+    Doughnut
   },
   props: {
     chartId: {
       type: String,
-      default: 'bar-chart'
+      default: 'doughnut-chart'
     },
     width: {
       type: Number,
@@ -36,35 +36,21 @@ export default defineComponent({
       type: String
     },
     styles: {
-      type: Object,
+      type: Object as PropType<Partial<CSSStyleDeclaration>>,
       default: () => {}
     },
     plugins: {
-      type: Object,
+      type: Object as PropType<PluginOptionsByType<'doughnut'>>,
       default: () => {}
     }
   },
   setup(props) {
     const chartData = {
-      labels: [
-        'January',
-        'February',
-        'March',
-        'April',
-        'May',
-        'June',
-        'July',
-        'August',
-        'September',
-        'October',
-        'November',
-        'December'
-      ],
+      labels: ['VueJs', 'EmberJs', 'ReactJs', 'AngularJs'],
       datasets: [
         {
-          label: 'Data One',
-          backgroundColor: '#f87979',
-          data: [40, 20, 12, 39, 10, 40, 39, 80, 40, 20, 12, 11]
+          backgroundColor: ['#41B883', '#E46651', '#00D8FF', '#DD1B16'],
+          data: [40, 20, 80, 10]
         }
       ]
     }
@@ -75,7 +61,7 @@ export default defineComponent({
     }
 
     return () =>
-      h(Bar, {
+      h(Doughnut, {
         chartData,
         chartOptions,
         chartId: props.chartId,
@@ -87,4 +73,3 @@ export default defineComponent({
       })
   }
 })
-</script>
