+++ sandboxes/radar/src/components/radarChart.ts
@@ -1,27 +1,27 @@
-<script>
-import { defineComponent, h } from 'vue'
+import { defineComponent, h, PropType } from 'vue'
 
-import { PolarArea } from 'vue-chartjs'
+import { Radar } from 'vue-chartjs'
 import {
   Chart as ChartJS,
   Title,
   Tooltip,
   Legend,
-  ArcElement,
-  RadialLinearScale
+  PointElement,
+  RadialLinearScale,
+  PluginOptionsByType
 } from 'chart.js'
 
-ChartJS.register(Title, Tooltip, Legend, ArcElement, RadialLinearScale)
+ChartJS.register(Title, Tooltip, Legend, PointElement, RadialLinearScale)
 
 export default defineComponent({
-  name: 'PolarAreaChart',
+  name: 'RadarChart',
   components: {
-    PolarArea
+    Radar
   },
   props: {
     chartId: {
       type: String,
-      default: 'polar-chart'
+      default: 'radar-chart'
     },
     width: {
       type: Number,
@@ -36,11 +36,11 @@ export default defineComponent({
       type: String
     },
     styles: {
-      type: Object,
+      type: Object as PropType<Partial<CSSStyleDeclaration>>,
       default: () => {}
     },
     plugins: {
-      type: Object,
+      type: Object as PropType<PluginOptionsByType<'radar'>>,
       default: () => {}
     }
   },
@@ -59,6 +59,7 @@ export default defineComponent({
         {
           label: 'My First dataset',
           backgroundColor: 'rgba(179,181,198,0.2)',
+          borderColor: 'rgba(179,181,198,1)',
           pointBackgroundColor: 'rgba(179,181,198,1)',
           pointBorderColor: '#fff',
           pointHoverBackgroundColor: '#fff',
@@ -68,6 +69,7 @@ export default defineComponent({
         {
           label: 'My Second dataset',
           backgroundColor: 'rgba(255,99,132,0.2)',
+          borderColor: 'rgba(255,99,132,1)',
           pointBackgroundColor: 'rgba(255,99,132,1)',
           pointBorderColor: '#fff',
           pointHoverBackgroundColor: '#fff',
@@ -83,7 +85,7 @@ export default defineComponent({
     }
 
     return () =>
-      h(PolarArea, {
+      h(Radar, {
         chartData,
         chartOptions,
         chartId: props.chartId,
@@ -95,4 +97,3 @@ export default defineComponent({
       })
   }
 })
-</script>
