+++ sandboxes/reactive/package.json
@@ -1,8 +1,9 @@
 {
   "name": "vue-chartjs-reactive-example",
   "dependencies": {
-    "vue": "^2.6.11",
+    "vue": "^3.2.31",
     "chart.js": "3.7.0",
-    "vue-chartjs": "3.5.1"
+    "vue-chartjs": "3.5.1",
+    "typescript": "^4.6.2"
   }
 }
