+++ sandboxes/scatter/src/components/scatterChart.ts
@@ -1,27 +1,27 @@
-<script>
-import { defineComponent, h } from 'vue'
-import { Bubble } from 'vue-chartjs'
+import { defineComponent, h, PropType } from 'vue'
 
+import { Scatter } from 'vue-chartjs'
 import {
   Chart as ChartJS,
   Title,
   Tooltip,
   Legend,
-  PointElement,
-  LinearScale
+  LineElement,
+  CategoryScale,
+  PluginOptionsByType
 } from 'chart.js'
 
-ChartJS.register(Title, Tooltip, Legend, PointElement, LinearScale)
+ChartJS.register(Title, Tooltip, Legend, LineElement, CategoryScale)
 
 export default defineComponent({
-  name: 'BubbleChart',
+  name: 'ScatterChart',
   components: {
-    Bubble
+    Scatter
   },
   props: {
     chartId: {
       type: String,
-      default: 'bubble-chart'
+      default: 'scatter-chart'
     },
     width: {
       type: Number,
@@ -36,11 +36,11 @@ export default defineComponent({
       type: String
     },
     styles: {
-      type: Object,
+      type: Object as PropType<Partial<CSSStyleDeclaration>>,
       default: () => {}
     },
     plugins: {
-      type: Object,
+      type: Object as PropType<PluginOptionsByType<'scatter'>>,
       default: () => {}
     }
   },
@@ -48,44 +48,58 @@ export default defineComponent({
     const chartData = {
       datasets: [
         {
-          label: 'Data One',
+          label: 'Scatter Dataset 1',
+          fill: false,
+          borderColor: '#f87979',
           backgroundColor: '#f87979',
           data: [
             {
-              x: 20,
-              y: 25,
-              r: 5
+              x: -2,
+              y: 4
             },
             {
-              x: 40,
-              y: 10,
-              r: 10
+              x: -1,
+              y: 1
             },
             {
-              x: 30,
-              y: 22,
-              r: 30
+              x: 0,
+              y: 0
+            },
+            {
+              x: 1,
+              y: 1
+            },
+            {
+              x: 2,
+              y: 4
             }
           ]
         },
         {
-          label: 'Data Two',
-          backgroundColor: '#7C8CF8',
+          label: 'Scatter Dataset 2',
+          fill: false,
+          borderColor: '#7acbf9',
+          backgroundColor: '#7acbf9',
           data: [
             {
-              x: 10,
-              y: 30,
-              r: 15
+              x: -2,
+              y: -4
             },
             {
-              x: 20,
-              y: 20,
-              r: 10
+              x: -1,
+              y: -1
             },
             {
-              x: 15,
-              y: 8,
-              r: 30
+              x: 0,
+              y: 1
+            },
+            {
+              x: 1,
+              y: -1
+            },
+            {
+              x: 2,
+              y: -4
             }
           ]
         }
@@ -98,7 +112,7 @@ export default defineComponent({
     }
 
     return () =>
-      h(Bubble, {
+      h(Scatter, {
         chartData,
         chartOptions,
         chartId: props.chartId,
@@ -110,4 +124,3 @@ export default defineComponent({
       })
   }
 })
-</script>
