+++ sandboxes/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "vue-chartjs": ["../src"]
+    }
+  },
+  "include": ["."]
+}
